update docs, move keyid to crypto module #105
