+++ dams/src/crypto.rs
@@ -2,15 +2,19 @@ use bytes::BytesMut;
 use rand::{CryptoRng, RngCore};
 use serde::{Deserialize, Serialize};
 
-use crate::{keys::KeyId, user::UserId};
+use crate::user::UserId;
 
 /// An export key is secure key material produced as client output from OPAQUE.
 ///
 /// This uses standardized naming, but is _not_ directly used as an encryption
-/// key in this system. Instead, it is used to derive a `MasterKey`.
+/// key in this system. Instead, the client uses it to derive a [`MasterKey`].
 ///
-/// Implementation note: this will be a wrapper around `opaque_ke`'s `export_key`
-/// field in the
+/// This key should not be stored or saved beyond the lifetime of a single
+/// authentication session.
+/// It should never be sent to the server.
+///
+/// Implementation note: this will be a wrapper around `opaque_ke`'s
+/// `export_key` field in the
 /// [ClientRegistrationFinishResult](opaque_ke::ClientRegistrationFinishResult)
 /// and corresponding registration result.
 #[allow(unused)]
@@ -25,7 +29,11 @@ impl ExportKey {
 }
 
 /// The master key is a default-length symmetric key for (TODO #107: encryption
-/// scheme). It is used by the user to securely encrypt their [`StorageKey`].
+/// scheme).
+///
+/// It is used by the client to securely encrypt their [`StorageKey`]. It should
+/// not be stored or saved beyond the lifetime of a single authentication
+/// session. It should never be sent to the server.
 #[allow(unused)]
 #[derive(Debug)]
 pub struct MasterKey;
@@ -41,11 +49,12 @@ impl MasterKey {
     }
 }
 
-/// A storage key is used to encrypt stored data. It is derived by and only
-/// known to a client.
-///
 /// A storage key is a default-length symmetric key for (TODO #107: encryption
-/// scheme).
+/// scheme) used to encrypt stored data.
+///
+/// It generated by the client and should never be revealed to the server.
+/// It should not be stored or saved beyond the lifetime of a single
+/// authentication session.
 #[allow(unused)]
 #[derive(Debug)]
 pub struct StorageKey;
@@ -66,9 +75,24 @@ impl StorageKey {
     }
 }
 
+/// Universally unique identifier for a [`Secret`].
+#[allow(unused)]
+#[derive(Debug, Serialize, Deserialize)]
+pub struct KeyId;
+
+#[allow(unused)]
+impl KeyId {
+    /// Generate a new, random `KeyId` for the given [`UserId`].
+    ///
+    /// This is called by the key server.
+    fn generate(rng: impl CryptoRng + RngCore, user_id: UserId) -> Self {
+        todo!()
+    }
+}
+
 /// An arbitrary secret.
 ///
-/// This is generated by and known only to the client.
+/// This is generated by the client and should never be revealed to the server.
 #[derive(Debug, Deserialize, Serialize)]
 pub struct Secret {
     material: BytesMut,
@@ -101,7 +125,7 @@ impl Default for Secret {
 }
 
 /// Context and application specific information used in various cryptographic
-/// protocols. This enum lists the specific contexts that arise throughout the
+/// protocols. This lists the specific contexts that arise throughout the
 /// system.
 ///
 /// Context is used in key derivation and authenticated encryption with
@@ -152,5 +176,12 @@ mod test {
         let thread_rng = rand::thread_rng();
         let user_id = UserId::default();
         let _secret = Secret::generate(thread_rng, 32, user_id, KeyId);
+    }
+
+    #[test]
+    #[should_panic(expected = "not yet implemented")]
+    fn key_id_generation_not_implemented() {
+        let thread_rng = rand::thread_rng();
+        let _key_id = KeyId::generate(thread_rng, UserId::default());
     }
 }
